+++ src/test/java/info/kfgodel/dyna/WorldTestContext.java
@@ -1,8 +1,8 @@
 package info.kfgodel.dyna;
 
 import ar.com.dgarcia.javaspec.api.contexts.TestContext;
-import info.kfgodel.dyna.api.Environment;
 import info.kfgodel.dyna.api.ObjectCreator;
+import info.kfgodel.dyna.api.environment.Environment;
 import info.kfgodel.dyna.testobjects.InterdependentTypeA;
 import info.kfgodel.dyna.testobjects.InterdependentTypeB;
